requirements.txt
accelerate==1.6.0
click==8.1.8
decorator==5.2.1
deepspeed==0.16.7
diffusers==0.34.0
einops==0.8.1
feedparser==6.0.11
flash_attn==2.3.4
huggingface-hub==0.31.1
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
json5==0.12.0
lazy_loader==0.4
lightning-utilities==0.14.3
matplotlib==3.10.1
matplotlib-inline==0.1.7
mmengine==0.10.7
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.5
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
pandas==2.2.3
peft==0.15.2
pillow==11.2.1
pytz==2025.2
PyYAML==6.0.2
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
six==1.17.0
timm==0.9.12
tokenizers==0.21.2
torch==2.7.0
torch-fidelity==0.3.0
torchmetrics==1.7.2
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.49.0
transformers-stream-generator==0.0.5
triton==3.3.0
xtuner==0.1.23
yarl==1.20.0

